% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getBG}
\alias{getBG}
\title{Get the statsus of using the background signature}
\usage{
getBG(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
the status of using the background signature
}
\description{
Get the statsus of using the background signature
}
