% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_G.R
\name{getCounts}
\alias{getCounts}
\title{Get the count data in a matrix}
\usage{
getCounts(object)
}
\arguments{
\item{object}{the MutationFeatureData class}
}
\value{
the count data in a matrix
}
\description{
Get the count data in a matrix
}
\examples{
print(getCounts(G))

}
