% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getExposures}
\alias{getExposures}
\title{Get a matrix of mutational exposures of signatures}
\usage{
getExposures(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
a matrix of mutational exposures of signatures
}
\description{
Get a matrix of mutational exposures of signatures
}
