% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getFeatures}
\alias{getFeatures}
\title{Get a vector of possible features}
\usage{
getFeatures(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
a vector of possible features
}
\description{
Get a vector of possible features
}
