% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_G.R
\name{getFeatureVec}
\alias{getFeatureVec}
\title{Get a matrix of feature vector list}
\usage{
getFeatureVec(object)
}
\arguments{
\item{object}{the MutationFeatureData class}
}
\value{
a matrix of feature vector list
}
\description{
Get a matrix of feature vector list
}
