% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getK}
\alias{getK}
\title{Get the number of signatures}
\usage{
getK(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
the number of signatures in \code{pmgetSignature} in \code{HiLDA}
}
\description{
Get the number of signatures
}
\examples{
print(getK(param))

}
