% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getLL}
\alias{getLL}
\title{Get the values of loglikelihood}
\usage{
getLL(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
likelihood values estimated by \code{pmgetSignature} in \code{HiLDA}
}
\description{
Get the values of loglikelihood
}
