% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_G.R
\name{getSamplelistG}
\alias{getSamplelistG}
\title{Get the sample list}
\usage{
getSamplelistG(object)
}
\arguments{
\item{object}{the MutationFeatureData class}
}
\value{
the sample list of named elements.
}
\description{
Get the sample list
}
