% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getSignatures}
\alias{getSignatures}
\title{Get an array of signature feature distributions}
\usage{
getSignatures(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
an array of signature feature distributions
}
\description{
Get an array of signature feature distributions
}
