% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_G.R
\name{getTranscription}
\alias{getTranscription}
\title{Get the statsus of specifying the transcription bias}
\usage{
getTranscription(object)
}
\arguments{
\item{object}{the MutationFeatureData class}
}
\value{
the status of specifying the transcription bias
}
\description{
Get the statsus of specifying the transcription bias
}
\examples{
print(getTranscription(G))

}
