% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_kth_fold.R
\name{select_kth_fold}
\alias{select_kth_fold}
\title{Output the training data or test data}
\usage{
select_kth_fold(inputG, k, f_s, folds, include = TRUE)
}
\arguments{
\item{inputG}{a MutationFeatureData S4 class output by the pmsignature.}

\item{k}{an integer number of the number of cross-validation folds.}

\item{f_s}{a primary key of combining the feature pattern and sample ID.}

\item{folds}{the assignment to each fold.}

\item{include}{a boolean indictor of whether to include kth fold or not.}
}
\value{
a MutationFeatureData S4 class of either include or exclude kth fold.
}
\description{
Output the training data or test data
}
\examples{

G_test <- select_kth_fold(G, k = 3, include=TRUE)

}
