% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_G.R
\name{splitG}
\alias{splitG}
\title{Output the maximum potential scale reduction statistic of all parameters
estimated}
\usage{
splitG(inputG, Kfold = 3)
}
\arguments{
\item{inputG}{a MutationFeatureData S4 class output by the pmsignature.}

\item{Kfold}{an integer number of the number of cross-validation folds.}
}
\value{
a matrix made of perplexity from the results of cross-validation.
}
\description{
Output the maximum potential scale reduction statistic of all parameters
estimated
}
\examples{

load(system.file("extdata/sample.rdata", package = "selectKSigs"))
G_split <- splitG(G, Kfold = 3)

}
