% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_PMSignature.R
\name{cv_PMSignature}
\alias{cv_PMSignature}
\title{Output the maximum potential scale reduction statistic of all parameters
estimated}
\usage{
cv_PMSignature(inputG, Kfold = 3, nRep = 3, Klimit = 8)
}
\arguments{
\item{train}{a MutationFeatureData S4 class output of training data.}

\item{test}{a MutationFeatureData S4 class output of test data.}

\item{param}{}
}
\value{
the likelihood of the test data
}
\description{
Output the maximum potential scale reduction statistic of all parameters
estimated
}
\examples{

res <- cv_PMSignature(G, Kfold = 3)

}
