% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_param.R
\name{getSamplelist}
\alias{getSamplelist}
\title{Get the sample list}
\usage{
getSamplelist(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of pmgetSignature)}
}
\value{
the sample list of named elements.
}
\description{
Get the sample list
}
